+++ packages/brick-container/serve/getProxy.js
@@ -5,6 +5,7 @@ import { getBrickPackages } from "@next-core/serve-helpers";
 import { getStoryboards } from "./utils/getStoryboards.js";
 import { fixV2Storyboard } from "./utils/fixV2Storyboard.js";
 import { injectIndexHtml } from "./utils/injectIndexHtml.js";
+import { getProcessedPublicDeps } from "./utils/getProcessedPublicDeps.js";
 import { concatBrickPackages } from "./utils/concatBrickPackages.js";
 
 const { safeDump, JSON_SCHEMA } = jsYaml;
@@ -24,13 +25,38 @@ export default function getProxy(env, getRawIndexHtml) {
     server,
   } = env;
   if (useRemote) {
+    const apiProxyOptions = getBasicProxyOptions(env, "api/");
+
     return [
       {
         ...getBasicProxyOptions(env, "api/websocket_service/"),
         ws: true,
       },
       {
-        ...getBasicProxyOptions(env, "api/"),
+        ...apiProxyOptions,
+        context: (pathname, req) => {
+          // DO NOT intercept SSE requests.
+          const matched =
+            pathname.startsWith(apiProxyOptions.context) &&
+            req.headers["accept"] === "text/event-stream";
+          return matched;
+        },
+        onProxyReq(proxyReq, req) {
+          // Reset the origin header to the remote server
+          if (req.headers["origin"] === `http://${host}:${port}`) {
+            proxyReq.setHeader("origin", server);
+          }
+        },
+      },
+      {
+        ...apiProxyOptions,
+        context: (pathname, req) => {
+          // Intercept requests other than  SSE.
+          const matched =
+            pathname.startsWith(apiProxyOptions.context) &&
+            req.headers["accept"] !== "text/event-stream";
+          return matched;
+        },
         selfHandleResponse: true,
         bypass(req) {
           const appId = getAppIdFromBootstrapPath(req.path);
@@ -233,16 +259,10 @@ export default function getProxy(env, getRawIndexHtml) {
                     )?.[1];
 
                     if (publicDeps && localBricks?.length) {
-                      try {
-                        const parsedPublicDeps = JSON.parse(publicDeps).filter(
-                          (item) =>
-                            !localBricks.includes(item.filePath.split("/")[1])
-                        );
-
-                        publicDeps = JSON.stringify(parsedPublicDeps);
-                      } catch (_err) {
-                        console.error(`JSON.parse() error: ${publicDeps}`);
-                      }
+                      publicDeps = getProcessedPublicDeps(
+                        publicDeps,
+                        localBricks
+                      );
                     }
 
                     const appRootTpl = content.match(
@@ -319,8 +339,28 @@ export default function getProxy(env, getRawIndexHtml) {
                     });
                   }
                   return injectIndexHtml(env, rawIndexHtml);
-                } else if (baseHref !== "/next/") {
-                  return content.replaceAll("/next/", baseHref);
+                } else {
+                  let htmlContent = content;
+
+                  if (baseHref !== "/next/") {
+                    htmlContent = htmlContent.replaceAll("/next/", baseHref);
+                  }
+
+                  const publicDeps = htmlContent.match(
+                    /\bw\.PUBLIC_DEPS\s*=\s*(\[[^;]*\])\s*;/
+                  )?.[1];
+
+                  if (publicDeps && localBricks?.length) {
+                    htmlContent = htmlContent.replace(
+                      /\bw\.PUBLIC_DEPS\s*=\s*\[[^;]*\]\s*;/,
+                      `w.PUBLIC_DEPS=${getProcessedPublicDeps(
+                        publicDeps,
+                        localBricks
+                      )};`
+                    );
+                  }
+
+                  return htmlContent;
                 }
               }
             }