+++ bricks/e2e/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.13](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.12...@next-bricks/e2e@0.3.13) (2024-05-09)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.3.12](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.11...@next-bricks/e2e@0.3.12) (2024-05-09)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.3.11](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.10...@next-bricks/e2e@0.3.11) (2024-05-07)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.3.10](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.9...@next-bricks/e2e@0.3.10) (2024-04-30)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.3.9](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.8...@next-bricks/e2e@0.3.9) (2024-04-22)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.3.8](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.7...@next-bricks/e2e@0.3.8) (2024-04-16)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
 ## [0.3.7](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.3.6...@next-bricks/e2e@0.3.7) (2024-04-14)
 
 **Note:** Version bump only for package @next-bricks/e2e