+++ packages/element/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.6](https://github.com/easyops-cn/next-core/compare/@next-core/element@1.2.5...@next-core/element@1.2.6) (2024-05-09)
+
+
+### Bug Fixes
+
+* use `Symbol.for` instead of `Symbol` ([0c62df8](https://github.com/easyops-cn/next-core/commit/0c62df862e466efb4120ef7e2445875e2393f888))
+
+
+
+
+
+## [1.2.5](https://github.com/easyops-cn/next-core/compare/@next-core/element@1.2.4...@next-core/element@1.2.5) (2024-05-09)
+
+**Note:** Version bump only for package @next-core/element
+
+
+
+
+
+## [1.2.4](https://github.com/easyops-cn/next-core/compare/@next-core/element@1.2.3...@next-core/element@1.2.4) (2024-04-16)
+
+**Note:** Version bump only for package @next-core/element
+
+
+
+
+
 ## [1.2.3](https://github.com/easyops-cn/next-core/compare/@next-core/element@1.2.2...@next-core/element@1.2.3) (2024-03-11)
 
 **Note:** Version bump only for package @next-core/element