+++ packages/eslint-config-next/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.18](https://github.com/easyops-cn/next-core/compare/@next-core/eslint-config-next@2.0.17...@next-core/eslint-config-next@2.0.18) (2024-04-16)
+
+**Note:** Version bump only for package @next-core/eslint-config-next
+
+
+
+
+
 ## [2.0.17](https://github.com/easyops-cn/next-core/compare/@next-core/eslint-config-next@2.0.16...@next-core/eslint-config-next@2.0.17) (2024-03-22)