+++ packages/pipes/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.19](https://github.com/easyops-cn/next-core/compare/@next-core/pipes@2.0.18...@next-core/pipes@2.0.19) (2024-05-09)
+
+**Note:** Version bump only for package @next-core/pipes
+
+
+
+
+
+## [2.0.18](https://github.com/easyops-cn/next-core/compare/@next-core/pipes@2.0.17...@next-core/pipes@2.0.18) (2024-04-16)
+
+**Note:** Version bump only for package @next-core/pipes
+
+
+
+
+
 ## [2.0.17](https://github.com/easyops-cn/next-core/compare/@next-core/pipes@2.0.16...@next-core/pipes@2.0.17) (2024-03-11)
 
 **Note:** Version bump only for package @next-core/pipes