+++ packages/yo/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.46](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.45...@next-core/yo@1.5.46) (2024-05-09)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.45](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.44...@next-core/yo@1.5.45) (2024-05-09)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.44](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.43...@next-core/yo@1.5.44) (2024-05-07)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
+## [1.5.43](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.42...@next-core/yo@1.5.43) (2024-04-16)
+
+**Note:** Version bump only for package @next-core/yo
+
+
+
+
+
 ## [1.5.42](https://github.com/easyops-cn/next-core/compare/@next-core/yo@1.5.41...@next-core/yo@1.5.42) (2024-04-08)
 
 **Note:** Version bump only for package @next-core/yo